@@ -112,7 +112,7 @@
         <!-- Maven build properties -->
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>1.7</java.version>
+        <java.version>1.8</java.version>
         <!-- Maven plugins versions -->
         <maven-enforcer-plugin.version>1.3.1</maven-enforcer-plugin.version>
         <sonar-maven-plugin.version>2.2</sonar-maven-plugin.version>
@@ -477,12 +477,16 @@
                         <configuration>
                             <rules>
                                 <requireMavenVersion>
-                                    <message>You are running an older version of Maven. JHipster requires at least Maven 3.0</message>
+                                    <message>You are running an older version of Maven. JHipster requires at least Maven
+                                        3.0
+                                    </message>
                                     <version>[3.0.0,)</version>
                                 </requireMavenVersion>
                                 <requireJavaVersion>
-                                    <message>You are running an older version of Java. JHipster requires at least JDK 1.7</message>
-                                    <version>[1.7.0,)</version>
+                                    <message>You are running an older version of Java. JHipster requires at least JDK
+                                        1.7
+                                    </message>
+                                    <version>[1.8.0,)</version>
                                 </requireJavaVersion>
                             </rules>
                         </configuration>
